+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- (back module) fix: should not flip the relation of `ignore_distance`. (Thanks to @wolgamir)
+- (command_warmup module) feature: add the `tag` structure for configuration node, and support the `fuji.command_warmup.bypass.<tag-name>` permission to bypass the warmup entry node for a player.